将所有dao中int64类型的成员改为int

+++ backend/go_gateway_demo/dao/service_grpc_rule.go
@@ -5,8 +5,8 @@ import (
 )
 
 type GrpcRule struct {
-	ID             int64  `json:"id" gorm:"primary_key"`
-	ServiceID      int64  `json:"service_id" gorm:"column:service_id" description:"服务ID"`
+	ID             int    `json:"id" gorm:"primary_key"`
+	ServiceID      int    `json:"service_id" gorm:"column:service_id" description:"服务ID"`
 	Port           int    `json:"port" gorm:"column:port" description:"端口"`
 	HeaderTransfor string `json:"header_transfor" gorm:"column:header_transfor" description:"header转换支持增加(add),删除（del），修改（edit）格式：add headername headervalue"`
 }
